Fix width of syntaxhighlighter class
When there is floating content on the right, use of "100%" erroneously
caused Firefox to extend the syntax highlighter block's width into the
floating content. And caused Chrome to shunt the content to below the
end of the floating block. Use of "auto" width instead sizes the width
of the syntaxhighlighter block properly, to stop before the edge of the
floating right content.

See also:
https://github.com/imagej/mediawiki/commit/350ca1599e7292522101b8c07424eca3f8e970fd